requirements.txt
easyocr
opencv-python
numpy
matplotlib
pillow
tensorflow
